Rename bitutil.h to bit_util.h
