+++ src/common/bit_util.h
